Add service layer with business logic for all entities

+++ src/main/java/org/nbu/repository/CashierRepo.java
@@ -11,5 +11,5 @@ import java.util.List;
 @Repository
 public interface CashierRepo extends JpaRepository<Cashier, Integer> {
     @Query("SELECT c FROM Cashier c WHERE c.store.id = :storeId AND c NOT IN (SELECT cr.cashier FROM CashRegister cr WHERE cr.cashier IS NOT NULL)")
-    List<Cashier> findByStoreIdWhenCashRegisterIsNull (@Param("storeId") Long storeId);
+    List<Cashier> findByStoreIdWhenCashRegisterIsNull (@Param("storeId") int storeId);
 }
